Address PR review feedback: remove redundant casts, clamp capacity, update docs
- Remove redundant int32() casts since defaultDeploymentCapacity is already int32
- Clamp defaulted capacity to MaxCapacity (step-aligned down) instead of
  failing, preventing unnecessary init failures in --no-prompt mode
- Update docs/examples to show capacity: 50

+++ cli/azd/extensions/azure.ai.agents/internal/cmd/init_foundry_resources_helpers.go
@@ -1175,7 +1175,7 @@ func resolveModelDeployments(
 		ModelName:    model.Name,
 		Options: &azdext.AiModelDeploymentOptions{
 			Locations: []string{location},
-			Capacity:  new(int32(defaultDeploymentCapacity)),
+			Capacity:  new(defaultDeploymentCapacity),
 		},
 		Quota: &azdext.QuotaCheckOptions{
 			MinRemainingCapacity: 1,
